+++ packages/iov-cli/src/cli.ts
@@ -58,7 +58,7 @@ export const main = (originalArgs: string[]): void => {
   console.log(colors.yellow("    - SetNameTx"));
   console.log(colors.yellow("    - TokenTicker"));
   console.log(colors.yellow("    - TransactionKind"));
-  console.log(colors.yellow("    - IovWriter"));
+  console.log(colors.yellow("    - MultiChainSigner"));
   console.log(colors.yellow("  * from @iov/crypto"));
   console.log(colors.yellow("    - Bip39"));
   console.log(colors.yellow("    - Ed25519"));
@@ -97,7 +97,7 @@ export const main = (originalArgs: string[]): void => {
       TokenTicker,
       TransactionKind,
       UserProfile,
-      IovWriter,
+      MultiChainSigner,
     } from "@iov/core";
     import { bnsCodec } from '@iov/bns';
     import { Bip39, Ed25519, Ed25519Keypair, Random, Sha256, Sha512 } from '@iov/crypto';