+++ packages/iov-core/types/multichainsigner.d.ts
@@ -1,12 +1,12 @@
 import { Address, BcpConnection, BcpTransactionResponse, ChainConnector, Nonce, UnsignedTransaction } from "@iov/bcp-types";
 import { KeyringEntryId, UserProfile } from "@iov/keycontrol";
 import { ChainId, PublicKeyBundle } from "@iov/tendermint-types";
-export declare class IovWriter {
+export declare class MultiChainSigner {
     readonly profile: UserProfile;
     private readonly knownChains;
     constructor(profile: UserProfile);
     chainIds(): ReadonlyArray<ChainId>;
-    reader(chainId: ChainId): BcpConnection;
+    connection(chainId: ChainId): BcpConnection;
     addChain(connector: ChainConnector): Promise<void>;
     keyToAddress(chainId: ChainId, key: PublicKeyBundle): Address;
     getNonce(chainId: ChainId, addr: Address): Promise<Nonce>;